Preventive Maintenance and Predictive Maintenance
Maintenance is an essential element for operating manufacturing equipment. The key difference between Preventive Maintenance and Predictive Maintenance is that preventive maintenance is scheduled on a regular basis, from time to time, while predictive maintenance is scheduled as needed, based on the asset’s condition. Both maintenance methods aim to increase the reliability of assets and reduce the likelihood of machine failure, thereby being cost-effective and reducing unplanned maintenance costs.

Preventive maintenance uses a schedule of inspections and tasks to find and fix small issues before they have a chance to develop into big problems like unexpected failures occurring in the future. It’s possible to stray over into over-maintenance, where doing more than that is needed.
Preventive maintenance comes with the challenge to balance the cost with returns.
Preventive maintenance can prevent equipment failures and extend the functional life of the assets and also keeps the production lines up and running, so as to be more profitable as well.


Predictive Maintenance tries to predict failure before it actually happens by monitoring the machine during normal operations by drawing real-time data collection and analysis of machine operation to identify issues at the nascent stage before they can interrupt production. Investing in a computerized maintenance management system will give the measurements and data required to make smart decisions and can develop a strong overall maintenance program to reduce unneeded maintenance tasks, minimize maintenance costs, and monitor the equipment and systems to keep the facility up and running.

In order to implement predictive maintenance, automated condition monitoring is need to be installed on the target equipment and machines. Brabo Edge platform, establishing a strong mesh of interconnected machines and equipment developed by Solulever, the company aims to achieve smart manufacturing for most manufacturers in the manufacturing industry.
https://solulever.com